requirements.txt
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.6.0
gplearn>=0.4.2
python-binance>=1.0.19
matplotlib>=3.7.0
ta-lib>=0.4.28